Move simplify from base.py to new algorithm sub-package
